# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Primera entrega del proyecto final del curso de Backend con Nodejs en Coderhouse. GitHub: https://github.com/FelixGonzalo/eCommerce_backend",
    "version": "0.1.0",
    "title": "eCommerce Backend "
  },
  "host": "localhost:8080",
  "basePath": "/api",
  "schemes": ["http"],
...
  "paths": {
    "/products": {
      "post": {
        "tags": ["products"],
        "summary": "add product",
        "description": "add product",
        "operationId": "addProduct",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "product object that needs to be added",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Product_variant1"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      },
      "get": {
        "tags": ["products"],
        "summary": "get all products",
        "description": "get all products",
        "operationId": "getAllProducts",
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "tags": ["products"],
        "summary": "get a product by id",
        "description": "get a product by id",
        "operationId": "getProductById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of product to return",
            "required": true,
            "type": "string",
            "format": "string"
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      },
      "put": {
        "tags": ["products"],
        "summary": "update a product by id",
        "description": "update a product by id",
        "operationId": "updateProduct",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of product to update",
            "required": true,
            "type": "string",
            "format": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "product object that needs to be updated",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Product_variant1"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      },
      "delete": {
        "tags": ["products"],
        "summary": "delete a product by id",
        "description": "delete a product by id",
        "operationId": "deleteProduct",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of product to delete",
            "required": true,
            "type": "string",
            "format": "string"
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Product_variant1": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string",
          "example": "TV"
        },
        "description": {
          "type": "string",
          "example": "TV ultra pro"
        },
        "code": {
          "type": "string",
          "example": "asdf32"
        },
        "thumbnail": {
          "type": "string",
          "example": "https://i.blogs.es/800731/captura-de-pantalla-2021-06-07-a-las-1.09.46/original.png"
        },
        "stock": {
          "type": "number",
          "example": 12
        },
        "price": {
          "type": "number",
          "example": 750.5
        }
      }
    }
  }
}